refactor(backend-cuda): split — extract BackendMoeFused (#8 Phase 5, final)
Final phase of the `cuda/mod.rs` split (#8). Move the last big
supertrait body out:

- `cuda/moe.rs` (~338 lines): `impl BackendMoeFused for CudaBackend` —
  `route_topk_softmax` + `try_gpu_route_topk_into_host` +
  `moe_align_block_size` + `moe_combine`. All four kernels run on
  `crate::ptx::MOE_*` PTX modules; no private helpers leak.

(The fused MoE Marlin GEMM phases — `moe_gemm_phase_batched` /
`moe_gemm_phase_vllm` — already moved with `BackendQuantMarlin` in
Phase 3. They're trait methods on `BackendQuantMarlin`, not
`BackendMoeFused`.)

After Phase 5: mod.rs **2302 → 1986 lines (-14%)**. Cumulative since
Phase 0: 4756 → 1986 (**-58%**). What's left in mod.rs is the core
`impl Backend for CudaBackend`, `CudaState` struct, process-global
stream + decode-state slots, and the `KvFp16` `BackendKvDtype` impl
— all genuinely "core CUDA" with no candidate for further split.
